Adding test to verify aliases
