+++ processing/src/main/java/org/apache/druid/math/expr/AliasExprMacro.java
@@ -1,0 +1,49 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.druid.math.expr;
+
+import java.util.List;
+
+/***
+ * Alias Expression macro to create an alias and delegate operations to the same base macro
+ */
+public class AliasExprMacro implements ExprMacroTable.ExprMacro
+{
+  private final ExprMacroTable.ExprMacro exprMacro;
+  private final String alias;
+
+  public AliasExprMacro(final ExprMacroTable.ExprMacro baseExprMacro, final String alias)
+  {
+    this.exprMacro = baseExprMacro;
+    this.alias = alias;
+  }
+
+  @Override
+  public Expr apply(List<Expr> args)
+  {
+    return exprMacro.apply(args);
+  }
+
+  @Override
+  public String name()
+  {
+    return alias;
+  }
+}
